fix in check_xcr0_ymm function

+++ common/sys/sysinfo.cpp
@@ -145,7 +145,7 @@ namespace embree
   __noinline bool check_xcr0_ymm() 
   {
 #if defined (__WIN32__)
-    int64 xcr0 = 0;
+    uint32_t xcr0 = 0;
 #if defined(__INTEL_COMPILER) 
     xcr0 = _xgetbv(0);
 #elif (defined(_MSC_VER) && (_MSC_FULL_VER >= 160040219)) // min VS2010 SP1 compiler is required
